Completed Menu options for EDIT MENU
The EDIT MENU contains the functions:
1) COPY
2) CUT
3) PASTE
4) CLEAR ALL
5) FIND
The implementation of COPY,CUT and PASTE was done using lambda function.
The implementation of CLEAR ALL required text_editor.delete("1.0",tk.END)
The implementation of find function was a bit lengthy. It was done using one function and two supporting functions.
A pop window with find and replace text inputs was made!
The buttons used the two supporting functions:
1) find_func
2) replace_func
For more information about find and replace functions, see the source code:
line 176 to 202 [line description currently acc to the time of commit]